update 调整 角色查询部门 返回值类型

+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysDeptMapper.java
@@ -28,7 +28,7 @@ public interface SysDeptMapper extends BaseMapperPlus<SysDept> {
      * @param deptCheckStrictly 部门树选择项是否关联显示
      * @return 选中部门列表
      */
-    List<Integer> selectDeptListByRoleId(@Param("roleId") Long roleId, @Param("deptCheckStrictly") boolean deptCheckStrictly);
+    List<Long> selectDeptListByRoleId(@Param("roleId") Long roleId, @Param("deptCheckStrictly") boolean deptCheckStrictly);
 
     /**
      * 修改子元素关系
